Redone installed packages tracking, was not working correctly
Package names only sometimes are equal with the
directory in feeds/$feedname/

One directory provides a single makefile, but
this can provide multiple packages.

To find the package names provided by the
symlinked directories in package/feeds/$feedname/
we now consult the feed's index and find all package
names that the Makefile provides.

All package names are listed in install_packages.sh
but only the first within a makefile is installed
when run, the others are commented out (but will
implicitly be installed anyway as they are in the
same Makefile)