save+tagversion: prevent trying to tag repos that are http(s)
These are likely not writable at all or require special auth (personal access token on github, e.g.), and are usually external public repos on which umbrella project related tags are unneeded at best, and usually unwanted at all.

So for now, we make the assumption non-http-URLs are developer's private repos that should receive tags, and http ones are not.