automatically chose gmake and make sure it is recent enough
openwrt 22.03 no longer works with outdated gnu make (such as present on macOS).
On macos, a recent version can be installed via brew, but it is installed as gmake
to prevent interference with the Apple installed make by default.
p44b now adapts to this and makes use of "gmake" when available.